requirements.txt
pymysql>=1.1.1
python-dotenv>=0.19.0
requests>=2.32.0